+++ Scripts/generate.sh
@@ -221,6 +221,8 @@ inject_v1_endpoint_paths() {
     case userGroupMembers(id: String)
     case userGroupMembersDelete(id: String)
 
+    case rolesSearch
+
 EOF
 
   cat > "$values_file" <<'EOF'
@@ -310,6 +312,8 @@ EOF
         case let .userGroup(id): return "usergroups/\(id)"
         case let .userGroupMembers(id): return "usergroups/\(id)/members"
         case let .userGroupMembersDelete(id): return "usergroups/\(id)/members/delete"
+
+        case .rolesSearch: return "roles/search"
 
 EOF